Data type by Char

+++ Part13/Part13/Program.cs
@@ -7,8 +7,8 @@ namespace variableProperties
         {
             Console.WriteLine("Data type : char");
             Console.WriteLine("Size :" + sizeof(char));
-            Console.WriteLine("Minimum Value :" + char.MinValue);
-            Console.WriteLine("Maximum Value :" + char.MaxValue);
+            Console.WriteLine("Minimum Value :" + (int)char.MinValue);
+            Console.WriteLine("Maximum Value :" + (int)char.MaxValue);
         }
     }
 }
